chore: WhereBuilder
