+++ packages/ui/src/elements/Table/index.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React from 'react'
 
 import type { Props } from './types'
